refactoring and organizing
